Organizando imports, contexto e configurando rota privada
