+++ src/app/pages/dashboard/Dashboard.tsx
@@ -1,19 +1,17 @@
+import { Padding } from "@mui/icons-material";
 import { Button } from "@mui/material";
-import { useAppDrawerContext } from "../../shared/context";
-import { useAppThemeContext } from "../../shared/context/ThemeContext";
-import { useContextUsuarioLogged } from "../../shared/hooks";
+import { useAppDrawerContext, useAppThemeContext, useContextUsuarioLogged } from "../../shared/context";
 
 export const Dashboard = () => {
 
     const { setToken } = useContextUsuarioLogged();
 
     const { toggleTheme } = useAppThemeContext();
-
     
     const { toggleDrawer } = useAppDrawerContext();
 
     return (
-        <div>
+        <div style={{padding: '20px'}}>
             <Button variant="outlined" onClick={() => toggleTheme()}>Toggle Theme</Button>
             <Button variant="outlined" onClick={() => toggleDrawer()}>Toggle Drawer</Button>
