+++ src/app/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './dashboard/Dashboard'
 export * from './login/Login'
+export * from './lista/Lista'
