+++ src/app/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './Rotas'
+export * from './Roteamento'
+export * from './RoteamentoPrivado'
