+++ src/app/shared/context/ThemeContext.tsx
@@ -33,9 +33,7 @@ export const AppThemeProvider = ({ children }: IThemeProviderProps) => {
     return (
         <AppThemeContext.Provider value={{ themeName, toggleTheme }}>
             <ThemeProvider theme={theme}>
-                <Box width='100vw' height='100vh' bgcolor={theme.palette.background.default}>
                     {children}
-                </Box>
             </ThemeProvider>
         </AppThemeContext.Provider>
     )
