+++ src/app/shared/context/index.ts
@@ -1,2 +1,3 @@
 export * from './UsuarioLogged'
 export * from './DrawerContext'
+export * from './ThemeContext'
