+++ src/app/shared/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useContextUsuarioLogged'
+export {}
