--- src/app/shared/hooks/useContextUsuarioLogged.ts
@@ -1,8 +1,0 @@
-import { useContext } from "react"
-import { UsuarioLoggedContext } from "../context"
-
-
-export const useContextUsuarioLogged = () => {
-    const context = useContext(UsuarioLoggedContext);
-    return context;
-}
