+++ src/app/shared/services/index.ts
@@ -1,1 +1,4 @@
-export {}
+export * from './api/ApiException'
+export * from './api/tarefas/TarefasService'
+export * from './api/usuario/UsuarioService'
+export * from './storage/useStorage'
